Improved: migrated the logic to schedule and update job in the component itself from action (#356)
